+++ src/Category/Changelog.md
@@ -8,6 +8,7 @@
 - Dropped PHP 8.3 support
 - Dropped Laravel 10 & 11 support
 - Added the `taxonsIn()` and `firstTaxonIn()` methods to the `HasTaxons` trait.
+- Added the `is_active` field to the `Taxon` model/table
 
 ## 5.x Series